Block Settings: Use Tag Processor to inject class name on wrapping element
When we introduced #42124 new block supports behavior we did so with a
PCRE replacement that opened the possibility for a few bugs related to
processing the HTML attributes. It was noted in that PR that this would
be a good candidate for the `WP_HTML_Tag_Processor`.

In this patch we're performing that replacement as follow-up work. This
should improve the reliability and hopefully the readability of what is
being done to the HTML as it renders.